refactor: group common logics in `__internal__`
