+++ integrations/react/src/__internal__/ActivityComponentType.tsx
