+++ integrations/react/src/__internal__/activity/ActivityProvider.tsx
