+++ integrations/react/src/__internal__/activity/index.ts
@@ -1,4 +1,3 @@
-export * from "./ActivityComponentType";
 export * from "./ActivityProvider";
 export * from "./makeActivityId";
 export * from "./makeStepId";
