+++ integrations/react/src/__internal__/activity/makeActivityId.ts
