+++ integrations/react/src/__internal__/activity/makeStepId.ts
