+++ integrations/react/src/__internal__/core/CoreProvider.tsx
