+++ integrations/react/src/__internal__/core/index.ts
