+++ integrations/react/src/__internal__/plugins/PluginsProvider.tsx
