+++ integrations/react/src/__internal__/plugins/index.ts
