+++ integrations/react/src/__internal__/shims/index.ts
