+++ integrations/react/src/__internal__/shims/useSyncExternalStore.ts
