+++ integrations/react/src/__internal__/stack/StackProvider.tsx
