+++ integrations/react/src/__internal__/stack/index.ts
