+++ integrations/react/src/__internal__/utils/index.ts
