+++ integrations/react/src/__internal__/utils/isBrowser.ts
