+++ integrations/react/src/__internal__/utils/isServer.ts
