+++ integrations/react/src/stable/BaseActivities.ts
@@ -1,6 +1,6 @@
 import type { ActivityRegisteredEvent } from "@stackflow/core";
 
-import type { ActivityComponentType } from "./activity";
+import type { ActivityComponentType } from "../__internal__/ActivityComponentType";
 
 export type BaseActivities = {
   [activityName: string]:
