+++ integrations/react/src/stable/index.ts
@@ -1,9 +1,9 @@
-export * from "./activity/ActivityComponentType";
-export * from "./activity/useActivity";
-export * from "./activity/useActivityParams";
-export * from "./stack/useStack";
+export * from "../__internal__/ActivityComponentType";
+export * from "../__internal__/activity/useActivity";
+export * from "../__internal__/activity/useActivityParams";
+export * from "../__internal__/stack/useStack";
+export * from "../__internal__/StackflowReactPlugin";
 export * from "./stackflow";
-export * from "./StackflowReactPlugin";
 export * from "./useActions";
 export * from "./useActiveEffect";
 export * from "./useEnterDoneEffect";
